mqsend: Set ulimit automatically during Open
When using mqsend in prod we need to set ulimit, mqsend.Open is the most
nature place to set that.
